+++ src/main/java/com/ariskourt/revolut/services/TransferValidationService.java
@@ -2,8 +2,10 @@ package com.ariskourt.revolut.services;
 
 import com.ariskourt.revolut.domain.BankAccount;
 
+import java.math.BigDecimal;
+
 public interface TransferValidationService {
 
-    void validateTransferDetails(BankAccount from, BankAccount to, Long transferAmount);
+    void validateTransferDetails(BankAccount from, BankAccount to, BigDecimal transferAmount);
 
 }
