+++ src/test/java/com/ariskourt/revolut/services/TransferValidationServiceTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -26,21 +27,21 @@ class TransferValidationServiceTest {
     @Test
     @DisplayName("When no from and to account are found validation fails")
     public void validateTransferDetails_WhenBothAccountsNotPresent_ExceptionIsThrown() {
-        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(null, null, 1000L));
+        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(null, null, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
     @DisplayName("When no from account is found validation fails")
     public void validateTransferDetails_WhenFromAccountNotPresent_ExceptionIsThrown() {
-        var from = createAccount(1000L);
-        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(from, null, 1000L));
+        var from = createAccount(1000.0);
+        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(from, null, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
     @DisplayName("When no to account is found validation fails")
     public void validateTransferDetails_WhenToAccountNotPresent_ExceptionIsThrown() {
-        var from = createAccount(1000L);
-        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(from, null, 1000L));
+        var from = createAccount(1000.0);
+        assertThrows(BankAccountNotFoundException.class, () -> validationService.validateTransferDetails(from, null, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
@@ -48,70 +49,70 @@ class TransferValidationServiceTest {
     public void validateTransferDetails_WhenAccountsAreTheSame_ExceptionIsThrown() {
         var createdAt = new Date();
         var id = UUID.randomUUID().toString();
-        var from = createAccount(id, 1000L, createdAt);
-        var to = createAccount(id, 1000L, createdAt);
-        assertThrows(SameAccountTransferException.class, () -> validationService.validateTransferDetails(from, to, 1000L));
+        var from = createAccount(id, 1000.0, createdAt);
+        var to = createAccount(id, 1000.0, createdAt);
+        assertThrows(SameAccountTransferException.class, () -> validationService.validateTransferDetails(from, to, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
     @DisplayName("When from account has no balance left, validation fails")
     public void validateTransferDetails_WhenFromAccountHasNoBalanceLeft_ExceptionIsThrown() {
-        var from = createAccount(0L);
-        var to = createAccount(1000L);
-        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, 1000L));
+        var from = createAccount(0.0);
+        var to = createAccount(1000.0);
+        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
     @DisplayName("When from account has negative balance, validation fails")
     public void validateTransferDetails_WhenAccountHasNegativeBalance_ExceptionIsThrown() {
-        var from = createAccount(-100L);
-        var to = createAccount(1000L);
-        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, 1000L));
+        var from = createAccount(-100.0);
+        var to = createAccount(1000.0);
+        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, BigDecimal.valueOf(1000.0)));
     }
 
     @Test
     @DisplayName("When from account has insufficient balance, validation fails")
     public void validateTransferDetails_WhenFromAccountHasInsufficientBalance_ExceptionIsThrown() {
-        var from = createAccount(500L);
-        var to = createAccount(1000L);
-        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, 2000L));
+        var from = createAccount(500.0);
+        var to = createAccount(1000.0);
+        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, BigDecimal.valueOf(2000.0)));
     }
 
     @Test
     @DisplayName("When from account has insufficient balance and transfer amount is negative, validation fails")
     public void validateTransferDetails_WhenFromAccountHasInsufficientBalanceNegativeTransferAmount_ExceptionIsThrown() {
-        var from = createAccount(500L);
-        var to = createAccount(1000L);
-        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, -2000L));
+        var from = createAccount(500.0);
+        var to = createAccount(1000.0);
+        assertThrows(InsufficientBalanceException.class, () -> validationService.validateTransferDetails(from, to, BigDecimal.valueOf(-2000.0)));
     }
 
     @Test
     @DisplayName("When both accounts are fine, validation succeeds")
     public void validateTransferDetails_WhenBothAccountsAreEligibleForTransfer_NoExceptionIsThrown() {
-        var from = createAccount(5000L);
-        var to = createAccount(1000L);
+        var from = createAccount(5000.0);
+        var to = createAccount(1000.0);
         try {
-            validationService.validateTransferDetails(from, to, 1000L);
+            validationService.validateTransferDetails(from, to, BigDecimal.valueOf(1000.0));
         } catch (Exception e) {
             fail();
         }
     }
 
-    private BankAccount createAccount(String id, Long balance, Date createdAt) {
+    private BankAccount createAccount(String id, Double balance, Date createdAt) {
         var account = new BankAccount();
         account.setId(id);
         account.setAccountHolder("John Doe");
-        account.setAccountBalance(balance);
+        account.setAccountBalance(BigDecimal.valueOf(balance));
         account.setCreatedAt(createdAt);
         account.setVersion(1);
         return account;
     }
 
-    private BankAccount createAccount(Long balance) {
+    private BankAccount createAccount(Double balance) {
         var account = new BankAccount();
         account.setId(UUID.randomUUID().toString());
         account.setAccountHolder("John Doe");
-        account.setAccountBalance(balance);
+        account.setAccountBalance(BigDecimal.valueOf(balance));
         account.setCreatedAt(new Date());
         account.setVersion(1);
         return account;
